authorization: Move ModulePattern, ResourcePattern to api
They will no longer be used in AccessRuleList going forward.
